(Apple/Metal) Combine various Metal files into one
