+++ gfx/common/metal/Shaders.metal
@@ -1,18 +1,26 @@
-//
-//  Shaders.metal
-//  MetalRenderer
-//
-//  Created by Stuart Carnie on 5/31/18.
-//  Copyright © 2018 Stuart Carnie. All rights reserved.
-//
+/*  RetroArch - A frontend for libretro.
+ *  Copyright (C) 2018      - Stuart Carnie
+ *  copyright (c) 2011-2021 - Daniel De Matteis
+ *
+ *  RetroArch is free software: you can redistribute it and/or modify it under the terms
+ *  of the GNU General Public License as published by the Free Software Found-
+ *  ation, either version 3 of the License, or (at your option) any later version.
+ *
+ *  RetroArch is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY;
+ *  without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR
+ *  PURPOSE.  See the GNU General Public License for more details.
+ *
+ *  You should have received a copy of the GNU General Public License along with RetroArch.
+ *  If not, see <http://www.gnu.org/licenses/>.
+ */
 
-// File for Metal kernel and shader functions
+/* File for Metal kernel and shader functions */
 
 #include <metal_stdlib>
 #include <simd/simd.h>
 
-// Including header shared between this Metal shader code and Swift/C code executing Metal API commands
-#import "ShaderTypes.h"
+/* Including header shared between this Metal shader code and Swift/C code executing Metal API commands */
+#import "metal_shader_types.h"
 
 using namespace metal;
