--- gfx/common/metal/TexturedView.h
@@ -1,23 +1,0 @@
-//
-// Created by Stuart Carnie on 6/16/18.
-//
-
-#import "Context.h"
-#import "View.h"
-
-@interface TexturedView : NSObject
-
-@property (nonatomic, readonly) RPixelFormat format;
-@property (nonatomic, readonly) RTextureFilter filter;
-@property (nonatomic, readwrite) BOOL visible;
-@property (nonatomic, readwrite) CGRect frame;
-@property (nonatomic, readwrite) CGSize size;
-@property (nonatomic, readonly) ViewDrawState drawState;
-
-- (instancetype)initWithDescriptor:(ViewDescriptor *)td context:(Context *)c;
-
-- (void)drawWithContext:(Context *)ctx;
-- (void)drawWithEncoder:(id<MTLRenderCommandEncoder>)rce;
-- (void)updateFrame:(void const *)src pitch:(NSUInteger)pitch;
-
-@end
