+++ gfx/common/metal/metal_shader_types.h
@@ -1,16 +1,25 @@
-//
-//  ShaderTypes.h
-//  MetalRenderer
-//
-//  Created by Stuart Carnie on 5/31/18.
-//  Copyright © 2018 Stuart Carnie. All rights reserved.
-//
+/*  RetroArch - A frontend for libretro.
+ *  Copyright (C) 2018      - Stuart Carnie
+ *  copyright (c) 2011-2021 - Daniel De Matteis
+ *
+ *  RetroArch is free software: you can redistribute it and/or modify it under the terms
+ *  of the GNU General Public License as published by the Free Software Found-
+ *  ation, either version 3 of the License, or (at your option) any later version.
+ *
+ *  RetroArch is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY;
+ *  without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR
+ *  PURPOSE.  See the GNU General Public License for more details.
+ *
+ *  You should have received a copy of the GNU General Public License along with RetroArch.
+ *  If not, see <http://www.gnu.org/licenses/>.
+ */
 
-//
-//  Header containing types and enum constants shared between Metal shaders and Swift/ObjC source
-//
-#ifndef ShaderTypes_h
-#define ShaderTypes_h
+/*
+ *  Header containing types and enum constants shared between Metal shaders and Swift/ObjC source
+ */
+
+#ifndef METAL_SHADER_TYPES_H
+#define METAL_SHADER_TYPES_H
 
 #ifdef __METAL_VERSION__
 #define NS_ENUM(_type, _name) enum _name : _type _name; enum _name : _type
@@ -88,4 +97,4 @@ typedef struct
    vector_float4 color;
 } FontFragmentIn;
 
-#endif /* ShaderTypes_h */
+#endif
